# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Catalog Service API
  version: 1.0.0
  description: Tenant-scoped category management API.
servers:
  - url: /api/v1/catalog
    description: Current catalog service host
...
paths:
...
  /categories:
    get:
      tags: [Categories]
      summary: List all categories
      description: Returns active and inactive categories for the authenticated user's tenant.
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        "200":
          description: Categories retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Categories]
      summary: Create a category
      security:
        - bearerAuth: []
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategory"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: categoryId
          in: query
          schema: { type: string }
      security:
        - bearerAuth: []
        - accessCookie: []
      responses:
        "200":
          description: Products retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Products]
      summary: Create a product with initial variants
      security:
        - bearerAuth: []
        - accessCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProduct" }
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessCookie:
      type: apiKey
      in: cookie
      name: access_token
...
  schemas:
    Category:
      type: object
      required: [id, tenantId, name, slug, displayOrder, isActive, createdAt, updatedAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string, maxLength: 100 }
        slug: { type: string, maxLength: 100 }
        parentId: { type: [string, "null"] }
        displayOrder: { type: integer, minimum: 0 }
        icon: { type: [string, "null"], maxLength: 50 }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateCategory:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        parentId: { type: [string, "null"] }
        displayOrder: { type: integer, minimum: 0, default: 0 }
        icon: { type: [string, "null"], maxLength: 50 }
...
    CategoryResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: "#/components/schemas/Category" }
        status: { type: integer, const: 200 }
    CategoryListResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: array
          items: { $ref: "#/components/schemas/Category" }
        status: { type: integer, const: 200 }
    Variant:
      type: object
      properties:
        id: { type: string }
        productId: { type: string }
        label: { type: string, maxLength: 50 }
        price: { type: string }
        isDefault: { type: boolean }
        isActive: { type: boolean }
        displayOrder: { type: integer }
    AddOn:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        price: { type: string }
        isActive: { type: boolean }
    Product:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        categoryId: { type: string }
        name: { type: string, maxLength: 150 }
        description: { type: [string, "null"] }
        imageUrl: { type: [string, "null"], format: uri }
        isVeg: { type: [boolean, "null"] }
        isActive: { type: boolean }
        displayOrder: { type: integer }
        attributes: { type: object }
        deletedAt: { type: [string, "null"], format: date-time }
        variants:
          type: array
          items: { $ref: "#/components/schemas/Variant" }
        addOns:
          type: array
          items: { $ref: "#/components/schemas/AddOn" }
    CreateProduct:
      type: object
      required: [name, categoryId, variants]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 150 }
        description: { type: [string, "null"] }
        imageUrl: { type: [string, "null"], format: uri }
        categoryId: { type: string }
        isVeg: { type: [boolean, "null"] }
        displayOrder: { type: integer, minimum: 0 }
        attributes: { type: object }
        addOnIds:
          type: array
          uniqueItems: true
          items: { type: string }
        variants:
          type: array
          minItems: 1
          items:
            type: object
            required: [label, price]
            properties:
              label: { type: string, maxLength: 50 }
              price: { type: number, minimum: 0 }
              isDefault: { type: boolean }
              displayOrder: { type: integer, minimum: 0 }
...
    ProductResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: "#/components/schemas/Product" }
        status: { type: integer }
    ProductListResponse:
      type: object
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: array
          items: { $ref: "#/components/schemas/Product" }
        status: { type: integer, const: 200 }
...
    ErrorResponse:
      type: object
      required: [success, message, status]
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        status: { type: integer }
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Insufficient permissions or no tenant
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Conflict:
      description: Category conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }